+++ src/azure-cli/HISTORY.rst
@@ -111,6 +111,8 @@ az webapp create : Fixing issue when running the command with --runtime
 **AKS**
 
 * Each cluster gets a separate service principal to improve isolation
+* Add --nodepool-tags to node pool when creating azure kubernetes cluster
+* Add --tags when adding or updating a nodepool to cluster
 
 **AppConfig**